[Widget] CR changes - Split logic from network and parsing, refactored tests and kept internal methods private, added copyright
Created a new object NetworkFeed to separate the networking part from the parsing part and moved the rssItems data to the WidgetFactory to use the object's methods just as helper functions. Made internal methods private again and just tested the main methods as black-box testing. Added copyright to new classes